// src/app/models/user-details.ts
// Define an interface for the user details
export interface UserDetails {
  name: string;
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  countryCode: string;
  countryName: string;
  userName: string;
  organizationId: string;
  password: string;
  pinCode: string;
  // Add other fields as needed...
}